2- changes in all of the app, changes type of price in database and models and some other little changes

+++ BaseProjectMvc_Net8/Models/ItemModel.cs
@@ -3,10 +3,13 @@
     public class ItemModel
     {
         public Guid Id { get; set; } = Guid.NewGuid();
+
         public int Product_Id { get; set; }
+
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
-        //public decimal Subtotal { get; set; }
+
+        public int Price { get; set; }
+
         public decimal Subtotal
         {
             get
@@ -14,8 +17,11 @@
                 return Price * Quantity;
             }
         }
+
         public string Product_Name { get; set; } = string.Empty;
+
         public int StockProducts { get; set; }
+
         public decimal Total { get; set; }
     }
 }
